removed partial agent (DQN and DDQNA), and refacto factorized one (D3QNAgent)
